==> jupyter/preparing_data/01 - Preparing Data.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Information" data-toc-modified-id="Information-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Information</a></span></li><li><span><a href="#Import-Packages" data-toc-modified-id="Import-Packages-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Import Packages</a></span></li><li><span><a href="#Teams" data-toc-modified-id="Teams-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Teams</a></span><ul class="toc-item"><li><span><a href="#Read-in-dataset" data-toc-modified-id="Read-in-dataset-3.1"><span class="toc-item-num">3.1&nbsp;&nbsp;</span>Read in dataset</a></span></li><li><span><a href="#Save-raw-table-to-html-for-blog" data-toc-modified-id="Save-raw-table-to-html-for-blog-3.2"><span class="toc-item-num">3.2&nbsp;&nbsp;</span>Save raw table to html for blog</a></span></li><li><span><a href="#Look-at-column-names" data-toc-modified-id="Look-at-column-names-3.3"><span class="toc-item-num">3.3&nbsp;&nbsp;</span>Look at column names</a></span></li><li><span><a href="#Remove-excess-columns" data-toc-modified-id="Remove-excess-columns-3.4"><span class="toc-item-num">3.4&nbsp;&nbsp;</span>Remove excess columns</a></span></li><li><span><a href="#Start-team-id-index-from-0" data-toc-modified-id="Start-team-id-index-from-0-3.5"><span class="toc-item-num">3.5&nbsp;&nbsp;</span>Start team id index from 0</a></span></li><li><span><a href="#Check-if-row-index-and-id-are-equal" data-toc-modified-id="Check-if-row-index-and-id-are-equal-3.6"><span class="toc-item-num">3.6&nbsp;&nbsp;</span>Check if row index and id are equal</a></span></li><li><span><a href="#Save-new-table-to-html-for-blog" data-toc-modified-id="Save-new-table-to-html-for-blog-3.7"><span class="toc-item-num">3.7&nbsp;&nbsp;</span>Save new table to html for blog</a></span></li><li><span><a href="#Save-table-to-csv-for-future-work" data-toc-modified-id="Save-table-to-csv-for-future-work-3.8"><span class="toc-item-num">3.8&nbsp;&nbsp;</span>Save table to csv for future work</a></span></li></ul></li><li><span><a href="#Players" data-toc-modified-id="Players-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Players</a></span><ul class="toc-item"><li><span><a href="#Read-in-data" data-toc-modified-id="Read-in-data-4.1"><span class="toc-item-num">4.1&nbsp;&nbsp;</span>Read in data</a></span></li><li><span><a href="#Save-raw-table-for-blog" data-toc-modified-id="Save-raw-table-for-blog-4.2"><span class="toc-item-num">4.2&nbsp;&nbsp;</span>Save raw table for blog</a></span></li><li><span><a href="#Look-at-column-names" data-toc-modified-id="Look-at-column-names-4.3"><span class="toc-item-num">4.3&nbsp;&nbsp;</span>Look at column names</a></span></li><li><span><a href="#Remove-excess-columns" data-toc-modified-id="Remove-excess-columns-4.4"><span class="toc-item-num">4.4&nbsp;&nbsp;</span>Remove excess columns</a></span></li><li><span><a href="#Give-fields-approprait-names" data-toc-modified-id="Give-fields-approprait-names-4.5"><span class="toc-item-num">4.5&nbsp;&nbsp;</span>Give fields approprait names</a></span></li><li><span><a href="#Restart-index-from-0" data-toc-modified-id="Restart-index-from-0-4.6"><span class="toc-item-num">4.6&nbsp;&nbsp;</span>Restart index from 0</a></span></li><li><span><a href="#Check-id-and-row-index-are-equal" data-toc-modified-id="Check-id-and-row-index-are-equal-4.7"><span class="toc-item-num">4.7&nbsp;&nbsp;</span>Check id and row index are equal</a></span></li><li><span><a href="#Save-table-to-html-for-blog" data-toc-modified-id="Save-table-to-html-for-blog-4.8"><span class="toc-item-num">4.8&nbsp;&nbsp;</span>Save table to html for blog</a></span></li><li><span><a href="#Save-table-to-csv-for-future-work" data-toc-modified-id="Save-table-to-csv-for-future-work-4.9"><span class="toc-item-num">4.9&nbsp;&nbsp;</span>Save table to csv for future work</a></span></li></ul></li><li><span><a href="#Fixtures" data-toc-modified-id="Fixtures-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>Fixtures</a></span><ul class="toc-item"><li><span><a href="#Read-in-data" data-toc-modified-id="Read-in-data-5.1"><span class="toc-item-num">5.1&nbsp;&nbsp;</span>Read in data</a></span></li><li><span><a href="#Save-raw-table-for-blog" data-toc-modified-id="Save-raw-table-for-blog-5.2"><span class="toc-item-num">5.2&nbsp;&nbsp;</span>Save raw table for blog</a></span></li><li><span><a href="#Look-at-column-names" data-toc-modified-id="Look-at-column-names-5.3"><span class="toc-item-num">5.3&nbsp;&nbsp;</span>Look at column names</a></span></li><li><span><a href="#Remove-excess-columns" data-toc-modified-id="Remove-excess-columns-5.4"><span class="toc-item-num">5.4&nbsp;&nbsp;</span>Remove excess columns</a></span></li><li><span><a href="#Start-index-from-0" data-toc-modified-id="Start-index-from-0-5.5"><span class="toc-item-num">5.5&nbsp;&nbsp;</span>Start index from 0</a></span></li><li><span><a href="#Check-that-id-and-row-index-match" data-toc-modified-id="Check-that-id-and-row-index-match-5.6"><span class="toc-item-num">5.6&nbsp;&nbsp;</span>Check that id and row index match</a></span></li><li><span><a href="#Save-table-as-html-for-blog" data-toc-modified-id="Save-table-as-html-for-blog-5.7"><span class="toc-item-num">5.7&nbsp;&nbsp;</span>Save table as html for blog</a></span></li><li><span><a href="#Save-table-to-csv-for-future-work" data-toc-modified-id="Save-table-to-csv-for-future-work-5.8"><span class="toc-item-num">5.8&nbsp;&nbsp;</span>Save table to csv for future work</a></span></li></ul></li><li><span><a href="#Player-Performance" data-toc-modified-id="Player-Performance-6"><span class="toc-item-num">6&nbsp;&nbsp;</span>Player Performance</a></span><ul class="toc-item"><li><span><a href="#Read-in-the-data" data-toc-modified-id="Read-in-the-data-6.1"><span class="toc-item-num">6.1&nbsp;&nbsp;</span>Read in the data</a></span></li><li><span><a href="#Look-at-columns" data-toc-modified-id="Look-at-columns-6.2"><span class="toc-item-num">6.2&nbsp;&nbsp;</span>Look at columns</a></span></li><li><span><a href="#Remove-excess-columns" data-toc-modified-id="Remove-excess-columns-6.3"><span class="toc-item-num">6.3&nbsp;&nbsp;</span>Remove excess columns</a></span></li><li><span><a href="#Start-indexing-from-0" data-toc-modified-id="Start-indexing-from-0-6.4"><span class="toc-item-num">6.4&nbsp;&nbsp;</span>Start indexing from 0</a></span></li><li><span><a href="#Save-new-table-to-html-for-blog" data-toc-modified-id="Save-new-table-to-html-for-blog-6.5"><span class="toc-item-num">6.5&nbsp;&nbsp;</span>Save new table to html for blog</a></span></li><li><span><a href="#Save-new-table-to-csv-for-future-work" data-toc-modified-id="Save-new-table-to-csv-for-future-work-6.6"><span class="toc-item-num">6.6&nbsp;&nbsp;</span>Save new table to csv for future work</a></span></li></ul></li></ul></div>

# FPL-ML-004: Preparing Data

## Information

In this notebook, the raw files will be read in and have some low-level preparation performed. This will include removing excess columns, making sure the LUTs (team, players and fixtures) ID match the row indexing.

## Import Packages

import json
import pandas as pd
import numpy as np
import collections

#pd.set_option('display.max_columns', None)

## Teams

### Read in dataset

with open('../../data/json/teams/teams.json', 'r') as f:
    teams = pd.DataFrame(json.load(f))

teams.head()

### Save raw table to html for blog

ncols = 4
teams.iloc[list(range(ncols)) + [len(teams) - 1]].transpose().to_html(
    '../../data/html/teams_raw.html', bold_rows=False)

### Look at column names

print(teams.columns)

### Remove excess columns

keep_these_columns = [
    'id', 'name', 'short_name', 'strength', 'strength_defence_home',
    'strength_attack_home', 'strength_overall_home', 'strength_defence_away',
    'strength_attack_away', 'strength_overall_away'
]

teams_cleaned = teams.copy()
teams_cleaned = teams_cleaned[keep_these_columns]
teams_cleaned.head()

### Start team id index from 0

teams_cleaned['id'] = teams_cleaned['id'] - 1
teams_cleaned.head(20)

### Check if row index and id are equal

If the row index and id are equal, then in future we can go directly to the correct fixture by using the row index rather than searching for the right id.

teams_cleaned['id'].equals(pd.Series(teams_cleaned.index))

### Save new table to html for blog

teams_cleaned.iloc[list(range(ncols)) +
                   [len(teams_cleaned) - 1]].transpose().to_html(
                       '../../data/html/teams_clean.html', bold_rows=False)

### Save table to csv for future work

teams_cleaned.to_csv(r'../../data/csv/teams.csv',
                     index=False,
                     index_label=False)

## Players

### Read in data

with open('../../data/json/elements/elements.json') as f:
    players = pd.DataFrame(json.load(f))

players.head()

### Save raw table for blog

ncols = 4
players.iloc[list(range(ncols)) + [len(players) - 1]].transpose().to_html(
    '../../data/html/players_raw.html', bold_rows=False)

### Look at column names

print(players.columns)

### Remove excess columns

keep_these_columns = [
    'id', 'first_name', 'second_name', 'web_name', 'element_type'
]

players_cleaned = players.copy()
players_cleaned = players_cleaned[keep_these_columns]
players_cleaned.head()

### Give fields approprait names

players_cleaned = players_cleaned.rename(columns={
    'web_name': 'display_name',
    'element_type': 'position'
})
players_cleaned.head()

### Restart index from 0

players_cleaned['id'] = players_cleaned['id'] - 1
players_cleaned['position'] = players_cleaned['position'] - 1
players_cleaned = players_cleaned.sort_values(by=['id'])
players_cleaned.head()

### Check id and row index are equal

teams_cleaned['id'].equals(pd.Series(teams_cleaned.index))

### Save table to html for blog

players_cleaned.iloc[list(range(ncols)) +
                     [len(players_cleaned) - 1]].transpose().to_html(
                         '../../data/html/players_clean.html', bold_rows=False)

### Save table to csv for future work

players_cleaned.to_csv(r'../../data/csv/players.csv',
                       index=False,
                       index_label=False)

## Fixtures
### Read in data

with open('../../data/json/fixtures/fixtures.json') as f:
    fixtures = pd.DataFrame(json.load(f))

fixtures.head()

### Save raw table for blog

ncols = 4
fixtures.iloc[list(range(ncols)) + [len(fixtures) - 1]].transpose().to_html(
    '../../data/html/fixtures_raw.html', bold_rows=False)

### Look at column names

print(fixtures.columns)

### Remove excess columns

keep_these_columns = [
    'id', 'kickoff_time', 'team_h', 'team_a', 'team_h_score', 'team_a_score',
    'team_h_difficulty', 'team_a_difficulty'
]

fixtures_cleaned = fixtures.copy()
fixtures_cleaned = fixtures_cleaned[keep_these_columns]
fixtures_cleaned.head()

### Start index from 0

fixtures_cleaned = fixtures_cleaned.sort_values(by=['id']).reset_index(
    drop=True)
fixtures_cleaned['id'] = fixtures_cleaned['id'] - 1
fixtures_cleaned['team_h'] = fixtures_cleaned['team_h'] - 1
fixtures_cleaned['team_a'] = fixtures_cleaned['team_a'] - 1
fixtures_cleaned.head()

### Check that id and row index match

fixtures_cleaned['id'].equals(pd.Series(fixtures_cleaned.index))

### Save table as html for blog

fixtures_cleaned.iloc[list(range(ncols)) +
                      [len(fixtures_cleaned) - 1]].transpose().to_html(
                          '../../data/html/fixtures_clean.html',
                          bold_rows=False)

### Save table to csv for future work

fixtures_cleaned.to_csv(r'../../data/csv/fixtures.csv',
                        index=False,
                        index_label=False)

## Player Performance

### Read in the data

nPlayers = 624
count = 0
gcn = False
data_dict = collections.OrderedDict()

for i in range(1, nPlayers + 1):

    # Open json file
    with open('../../data/json/element-summary/' + str(i) + '.json') as f:
        es = json.load(f)

    # Number of fixtures for this player
    nFixtures = len(es["history"])

    # Get column names. only do this once
    if gcn:
        cn = es["history"][0].keys()
        gcn = False

    # Loop through this players fixtures
    for m in range(0, nFixtures):
        a = es["history"][m]
        data_dict[count] = a
        count = count + 1

player_performance = pd.DataFrame.from_dict(data_dict, orient="index")

player_performance.head()

print(player_performance.shape)

### Look at columns

col_names = list(player_performance.columns)
print(col_names)

### Remove excess columns

player_performance_cleaned = player_performance.copy()

# Remove excess columns
player_performance_cleaned = player_performance_cleaned.drop(
    columns=['id', 'kickoff_time_formatted'])

# Rename element column to id
player_performance_cleaned = player_performance_cleaned.rename(
    columns={'element': 'id'})

# Convert kickoff time to datetime
player_performance_cleaned['kickoff_time'] = pd.to_datetime(
    player_performance_cleaned['kickoff_time'], utc=True)

# Sort rows first by player id, then by kickoff time
player_performance_cleaned = player_performance_cleaned.sort_values(
    by=['id', 'kickoff_time']).reset_index(drop=True)

player_performance_cleaned.head()

### Start indexing from 0

player_performance_cleaned['id'] = player_performance_cleaned['id'] - 1

player_performance_cleaned[
    'opponent_team'] = player_performance_cleaned['opponent_team'] - 1

player_performance_cleaned[
    'fixture'] = player_performance_cleaned['fixture'] - 1

player_performance_cleaned.head()

### Save new table to html for blog

ncols = 4
player_performance_cleaned.iloc[list(range(ncols)) +
                      [len(player_performance_cleaned) - 1]].transpose().to_html(
                          '../../data/html/player_performance_clean.html',
                          bold_rows=False)

### Save new table to csv for future work

player_performance_cleaned.to_csv(r'../../data/csv/player_performance.csv',
                        index=False,
                        index_label=False)
