==> jupyter/preparing_data/02 - Adding Team And Position.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Import-packages" data-toc-modified-id="Import-packages-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Import packages</a></span></li><li><span><a href="#Read-in-data" data-toc-modified-id="Read-in-data-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Read in data</a></span></li><li><span><a href="#Get-additional-player-information" data-toc-modified-id="Get-additional-player-information-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Get additional player information</a></span></li><li><span><a href="#Save-table-as-csv" data-toc-modified-id="Save-table-as-csv-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Save table as csv</a></span></li></ul></div>

# Add Team and Position

## Import packages

import pandas as pd
import numpy as np

## Read in data

player_performance = pd.read_csv('../../data/csv/player_performance.csv')
player_performance.head()

fixtures = pd.read_csv('../../data/csv/fixtures.csv')
fixtures.head()

players = pd.read_csv('../../data/csv/players.csv')
players.head()

## Get additional player information

# Number or rows in player_performance table
nrows = len(player_performance)

# Empty team array
team = np.zeros(nrows, dtype='int')
position = np.zeros(nrows, dtype='int')

# Loop through rows
for i in range(nrows):

    # Get team id
    current_fixture = player_performance.loc[i, 'fixture']
    if player_performance.loc[i, 'was_home']:
        team[i] = fixtures.loc[current_fixture, 'team_h']
    else:
        team[i] = fixtures.loc[current_fixture, 'team_a']

    # Get position
    current_id = player_performance.loc[i, 'id']
    position[i] = players.loc[current_id, 'position']

player_performance['team'] = pd.Series(team)
player_performance['position'] = pd.Series(position)
player_performance.head()

## Save table as csv

player_performance.to_csv(r'../../data/csv/player_performance_01.csv',
                          index=False,
                          index_label=False)
